+++ tests/testthat/test-get-data.R
@@ -77,31 +77,7 @@ test_that("pxlib_get_data reads english data with BLOBs correctly", {
   )
 })
 
-# Test case 4: An English files with BLOB/Memo fields
-test_that("pxlib_get_data reads english data with BLOBs correctly", {
-  # Path to the test database
-  db_path <- system.file("extdata", "mushrooms.db", package = "Rparadox")
-
-  # --- Action: Open the file and get data ---
-  px_doc <- pxlib_open_file(db_path)
-  data_tbl <- pxlib_get_data(px_doc)
-  pxlib_close_file(px_doc)
-
-  # --- Assertions ---
-  # 1. Check that the result is a tibble
-  expect_s3_class(data_tbl, "tbl_df")
-
-  # 2. Compare the result with its corresponding reference file
-  ref_path <- test_path("ref_mushrooms.rds")
-  expect_identical(
-    object = data_tbl,
-    expected = readRDS(ref_path),
-    label = "Data loaded from mushrooms.db",
-    expected.label = "Reference data from ref_mushrooms.rds"
-  )
-})
-
-# Test case 5: An Russian file with empty doscodepage
+# Test case 4: An Russian file with empty doscodepage
 test_that("pxlib_get_data reads Russian file with empty doscodepage correctly", {
   # Path to the test database
   db_path <- system.file("extdata", "of.db", package = "Rparadox")
@@ -125,7 +101,7 @@ test_that("pxlib_get_data reads Russian file with empty doscodepage correctly", 
   )
 })
 
-# Test case 6: An Russian file with doscodepage CP866
+# Test case 5: An Russian file with doscodepage CP866
 test_that("pxlib_get_data reads Russian file with doscodepage CP866 correctly", {
   # Path to the test database
   db_path <- system.file("extdata", "of_cp866.db", package = "Rparadox")
@@ -149,7 +125,7 @@ test_that("pxlib_get_data reads Russian file with doscodepage CP866 correctly", 
   )
 })
 
-# Test case 7: An empty file
+# Test case 6: An empty file
 test_that("pxlib_get_data handles an empty table gracefully", {
   # Path to the test database
   db_path <- system.file("extdata", "empty.db", package = "Rparadox")
@@ -174,7 +150,7 @@ test_that("pxlib_get_data handles an empty table gracefully", {
   )
 })
 
-# Test case 8: incorrect argument
+# Test case 7: incorrect argument
 test_that("pxlib_get_data validates input correctly", {
   expect_error(pxlib_get_data("invalid"), "class 'pxdoc_t'")
   expect_error(pxlib_get_data(123), "class 'pxdoc_t'")